refactor(frontend): domain 層導入と ReceiptTray 整理（#104-3）
按分業務ルールを domain/settlement/ に集約し、ReceiptTray のロジックを
useReceiptTrayController へ抽出。feature 間依存ルールと Mapper 方針を規約に追記。
